+++ src/backend/guild-server/src/main/java/com/asyncgate/guild_server/repository/DirectMemberRepository.java
@@ -6,4 +6,6 @@ import java.util.List;
 
 public interface DirectMemberRepository {
     void saveAll(List<DirectMember> directMembers);
+
+    List<DirectMember> getDirectMessageList(String currentUserId);
 }
